reorganize k8s resources
